+++ CHANGELOG.md
@@ -5,6 +5,7 @@ to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- bdbe6cb WIP: ChangelogChecker
 - 2d5fe52 WIP: IncreasedChecker Fix
 - 65f3836 Refactor commits
 - 0634b57 Update Increaded Checker and Defaultchecker